+++ src/helpers/constants/Game.js
@@ -19,6 +19,7 @@ exports.turnNightActionsOrder = [
     { source: "all", action: "elect-sheriff", isFirstNightOnly: true },
     { source: "cupid", action: "charm", isFirstNightOnly: true },
     { source: "lovers", action: "meet-each-other", isFirstNightOnly: true },
+    { source: "wild-child", action: "choose-model", isFirstNightOnly: true },
     { source: "seer", action: "look" },
     { source: "two-sisters", action: "meet-each-other" },
     { source: "three-brothers", action: "meet-each-other" },