+++ src/helpers/constants/Role.js
@@ -11,6 +11,7 @@ exports.roleNames = [
     "cupid",
     "two-sisters",
     "three-brothers",
+    "wild-child",
 ];
 
 exports.groupNames = ["villagers", "werewolves", "lovers"];
@@ -27,5 +28,6 @@ exports.roles = [
     { name: "cupid", group: "villagers", maxInGame: 1, powers: [{ name: "charm" }] },
     { name: "two-sisters", group: "villagers", minInGame: 2, maxInGame: 2, recommendedMinPlayers: 12 },
     { name: "three-brothers", group: "villagers", minInGame: 3, maxInGame: 3, recommendedMinPlayers: 15 },
+    { name: "wild-child", group: "villagers", maxInGame: 1 },
     { name: "werewolf", group: "werewolves", maxInGame: 4, powers: [{ name: "eat" }] },
 ];